crab rave only works in random

+++ cogs/fun-commands.py
@@ -1,4 +1,6 @@
 import io
+from os import getenv
+
 import aiohttp
 import discord
 from discord.ext import commands
@@ -8,18 +10,22 @@ from urllib import parse
 class FunCommands(commands.Cog, name="Fun Commands"):
     def __init__(self, bot):
         self.bot: commands.Bot = bot
+        self.random_channel = int(getenv("CHANNEL_RANDOM", 689534362760642676))
 
     @commands.command(aliases=['crabrave', 'crab_rave', 'crab-rave'], hidden=True)
     async def crab(self, ctx, *, text):
-        await ctx.message.delete()
-        url = f'https://adventurous-damselfly.glitch.me/video/{parse.quote(text)}.mp4?style=classic'
-        async with ctx.channel.typing():
-            async with aiohttp.ClientSession() as session:
-                async with session.get(url) as resp:
-                    if resp.status != 200:
-                        return await ctx.send('Could not download file...')
-                    data = io.BytesIO(await resp.read())
-                    await ctx.send(file=discord.File(data, f'{text}.mp4'))
+        if ctx.channel.id == self.random_channel:
+            await ctx.message.delete()
+            url = f'https://adventurous-damselfly.glitch.me/video/{parse.quote(text)}.mp4?style=classic'
+            async with ctx.channel.typing():
+                async with aiohttp.ClientSession() as session:
+                    async with session.get(url) as resp:
+                        if resp.status != 200:
+                            return await ctx.send('Could not download file...')
+                        data = io.BytesIO(await resp.read())
+                        await ctx.send(file=discord.File(data, f'{text}.mp4'))
+        else:
+            await ctx.send("Sorry, please do that in #random")
 
 
 def setup(bot):
